Phone number long

+++ app/src/main/java/com/example/parkinglot/User.kt
@@ -9,7 +9,7 @@ data class User(
     var firstname: String? = "",
     var gender: Int? = 0,
     var lastname: String? = "",
-    var phone_number: Int? = 0,
+    var phone_number: Long? = 0,
     var username: String? = "",
     var password: String? = "",
     var saldo: Int? = 0
